+++ src/MediawikiSession.php
@@ -4,6 +4,7 @@ namespace Mediawiki\Api;
 
 use Psr\Log\LoggerAwareInterface;
 use Psr\Log\LoggerInterface;
+use Psr\Log\LogLevel;
 
 /**
  * @since 0.1
@@ -76,6 +77,8 @@ class MediawikiSession implements LoggerAwareInterface {
 		// If we don't already have the token that we want
 		if( !array_key_exists( $type, $this->tokens ) ) {
 
+			$this->log( LogLevel::DEBUG, 'Getting fresh token', array( 'type' => $type ) );
+
 			// If we know that we don't have the new module mw<1.25
 			if( $this->usePre125TokensModule ) {
 				return $this->reallyGetPre125Token( $type );
@@ -111,6 +114,7 @@ class MediawikiSession implements LoggerAwareInterface {
 		if( array_key_exists( 'warnings', $result ) && array_key_exists( 'query', $result['warnings'] ) &&
 			strstr( $result['warnings']['query']['*'], "Unrecognized value for parameter 'meta': tokens" ) ) {
 			$this->usePre125TokensModule = true;
+			$this->log( LogLevel::DEBUG, 'Falling back to pre 1.25 token system' );
 			$this->tokens[$type] = $this->reallyGetPre125Token( $type );
 		} else {
 			$this->tokens[$type] = array_pop( $result['query']['tokens'] );
@@ -165,6 +169,7 @@ class MediawikiSession implements LoggerAwareInterface {
 	 * @since 0.2
 	 */
 	public function clearTokens() {
+		$this->log( LogLevel::DEBUG, 'Clearing session tokens', array( 'tokens' => $this->tokens ) );
 		$this->tokens = array();
 	}